fix: hold a map of pointers to processes to use reassigned values outside of process.go scope
This is necessary for the DeadCh to be listened outside of `process.go`
scope.
In  `ProcessStartAction` we reassign a value to the channel.
If we don't use pointers we can't get the value of the new channel.
To keep everything safe and to don't be tempted to directly modify
the struct `Process`, when we ask for a Program Process list, we return
a copy of the map holding struct and not pointers to struct.