fix: let the system release stdout and stderr manually
When testing with a large number of processes,
for some of them if try to close the file descriptors we get an error
saying that they have already been closed.
The [official documentation](https://golang.org/pkg/os/exec/#Cmd.Wait)
of (*Cmd).Wait says that it automatically releases all
associated resources.